fix(core): split observability utils to prevent AsyncLocalStorage in browser bundles
The playground was failing to load with 'AsyncLocalStorage is not a constructor'
because importing @mastra/core/observability caused the Node.js-only
AsyncLocalStorage API to be evaluated in the browser.

Split observability/utils.ts into:
- utils.ts: Browser-safe functions (getOrCreateSpan, getRootExportSpan, getEntityTypeForSpan)
- context-storage.ts: Server-only functions that use AsyncLocalStorage
  (getCurrentSpan, executeWithContext, executeWithContextSync)

Internal core imports now use context-storage.ts directly, while the public
observability index only exports browser-safe utilities.
