+++ packages/core/src/llm/model/model.loop.ts
@@ -7,7 +7,7 @@ import { loop } from '../../loop';
 import type { LoopOptions } from '../../loop/types';
 import type { Mastra } from '../../mastra';
 import { SpanType, resolveObservabilityContext } from '../../observability';
-import { executeWithContextSync } from '../../observability/utils';
+import { executeWithContextSync } from '../../observability/context-storage';
 import type { MastraModelOutput } from '../../stream/base/output';
 import type { ModelManagerModelConfig } from '../../stream/types';
 import { delay } from '../../utils';
