+++ packages/core/src/llm/model/model.ts
@@ -22,7 +22,7 @@ import { MastraBase } from '../../base';
 import { MastraError, ErrorDomain, ErrorCategory } from '../../error';
 import type { Mastra } from '../../mastra';
 import { SpanType, resolveObservabilityContext } from '../../observability';
-import { executeWithContext, executeWithContextSync } from '../../observability/utils';
+import { executeWithContext, executeWithContextSync } from '../../observability/context-storage';
 import { toStandardSchema, standardSchemaToJSONSchema, isStandardSchemaWithJSON } from '../../schema';
 import type { ZodSchema } from '../../schema';
 import { convertV4Usage } from '../../stream/aisdk/v4/usage';
