+++ packages/core/src/tools/tool-builder/builder.ts
@@ -14,15 +14,9 @@ import { z } from 'zod/v4';
 import { MastraBase } from '../../base';
 import { ErrorCategory, MastraError, ErrorDomain } from '../../error';
 import type { Mastra } from '../../mastra';
-import {
-  SpanType,
-  wrapMastra,
-  executeWithContext,
-  EntityType,
-  getOrCreateSpan,
-  createObservabilityContext,
-} from '../../observability';
+import { SpanType, wrapMastra, EntityType, getOrCreateSpan, createObservabilityContext } from '../../observability';
 import type { AnySpan } from '../../observability';
+import { executeWithContext } from '../../observability/context-storage';
 import { RequestContext } from '../../request-context';
 import { isStandardSchemaWithJSON, toStandardSchema, standardSchemaToJSONSchema } from '../../schema';
 import { isVercelTool, isProviderDefinedTool } from '../../tools/toolchecks';
