+++ packages/core/src/workflows/evented/step-executor.ts
@@ -8,7 +8,7 @@ import { RegisteredLogger } from '../../logger';
 import type { Mastra } from '../../mastra';
 import type { TracingContext } from '../../observability';
 import { createObservabilityContext } from '../../observability';
-import { executeWithContext } from '../../observability/utils';
+import { executeWithContext } from '../../observability/context-storage';
 import { ToolStream } from '../../tools/stream';
 import { PUBSUB_SYMBOL, STREAM_FORMAT_SYMBOL } from '../constants';
 import { getStepResult } from '../step';
