+++ packages/core/src/workflows/handlers/step.ts
@@ -8,11 +8,11 @@ import {
   EntityType,
   SpanType,
   wrapMastra,
-  executeWithContext,
   createObservabilityContext,
   resolveObservabilityContext,
 } from '../../observability';
 import type { ObservabilityContext, Span } from '../../observability';
+import { executeWithContext } from '../../observability/context-storage';
 import { ToolStream } from '../../tools/stream';
 import type { DynamicArgument } from '../../types';
 import { PUBSUB_SYMBOL, STREAM_FORMAT_SYMBOL } from '../constants';
